Move all impl TryFrom<SqliteRow> for {...} into that type's file.
