+++ backend/src/task.rs
@@ -1,11 +1,14 @@
 use std::cmp::max;
 
+use sqlx::sqlite::SqliteRow;
+use sqlx::Row;
+
 use chrono::{
     NaiveDate,
     Duration,
 };
 
-use crate::utils::today_date;
+use crate::utils::{today_date,parse_date};
 use crate::due_date::DueDate;
 
 pub type Id = Option<u32>;
@@ -66,10 +69,43 @@ impl std::default::Default for Task {
     }
 }
 
+impl TryFrom<SqliteRow> for Task {
+    type Error = anyhow::Error;
+
+    fn try_from(row: SqliteRow) -> Result<Self, Self::Error> {
+        Ok(Task {
+            category:                         row.get::<String, &str>("Category"),
+            finished:                         row.get::<bool,   &str>("Finished"),
+            name:                             row.get::<String, &str>("Name"),
+            _time_budgeted: Duration::minutes(row.get::<i64,    &str>("Budget")),
+            time_needed:    Duration::minutes(row.get::<i64,    &str>("Time")),
+            time_used:      Duration::minutes(row.get::<i64,    &str>("Used")),
+            next_action_date:     parse_date(&row.get::<String, &str>("NextAction"))?,
+            due_date:                         row.get::<String, &str>("DueDate").try_into()?,
+            notes:                            row.get::<String, &str>("Notes"),
+            id:                               row.get::<Option<u32>, &str>("TaskID"),
+            date_added:           parse_date(&row.get::<String, &str>("DateAdded"))?,
+            links:                            Vec::new(),
+        })
+    }
+}
+
 /// Stores the data for a hyperlink. This is a thin data class
 #[derive(Clone, Debug, Serialize, Deserialize, Hash, PartialEq, Eq, Default)]
 pub struct Hyperlink {
     pub url:     String,
     pub display: String,
     pub id:      usize,
+}
+
+impl TryFrom<SqliteRow> for Hyperlink {
+    type Error = std::convert::Infallible;
+
+    fn try_from(row: SqliteRow) -> Result<Self, Self::Error> {
+        Ok(Hyperlink {
+            url:     row.get::<String, &str>("Url"),
+            display: row.get::<String, &str>("Display"),
+            id:      row.get::<u32, &str>("rowid") as usize,
+        })
+    }
 }
